+++ src/OpenSleigh.Transport.Kafka/IBusConfiguratorExtensions.cs
@@ -47,7 +47,7 @@ namespace OpenSleigh.Transport.Kafka
                 var builder = ctx.GetRequiredService<ProducerBuilder<string, byte[]>>();
                 return builder.Build();
             });
-            busConfigurator.Services.AddTransient<IKafkaPublisherExecutor, KafkaPublisherExecutor>();
+            busConfigurator.Services.AddTransient<IKafkaPublisherExecutor, KafkaPublisher>();
             busConfigurator.Services.AddTransient<IPublisher, KafkaPublisher>();
 
             busConfigurator.Services.AddTransient<IMessageParser, MessageParser>();
